requirements.txt
asgiref==3.6.0
Django==4.1
django-browser-reload==1.7.0
sqlparse==0.4.3
django-cors-headers==3.14.0
djangorestframework==3.14.0
